% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dafnia_functions.R
\name{aggregate_result}
\alias{aggregate_result}
\title{aggregates a column in a dataframe by factors in other columns}
\usage{
aggregate_result(df, out, by)
}
\arguments{
\item{df}{a dataframe with columns \code{out} and elements of \code{by}}

\item{out}{a column name (string) in dataframe \code{df} that you
want to aggregate/summarize}

\item{by}{a vector of column names (strings) that you want to aggregate by}
}
\value{
a dataframe with the statistics of the aggregate
}
\description{
aggregates a column in a dataframe by factors in other columns
}
