% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This says hello to any person!}
\usage{
hello(name = "William")
}
\arguments{
\item{name}{A string of characters. Preferentially a name of a person. (Default 'William')}
}
\value{
Prints Hello name to screen
}
\description{
Here jkh
}
\examples{
hello('Malie')
}
