% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{redplot}
\alias{redplot}
\title{This function creates a plot with red dots}
\usage{
redplot(x, y, col = "red")
}
\arguments{
\item{x}{is a vector or dataframe}

\item{y}{is a vector or dataframe}
}
\value{
Creates a plot with red dots
}
\description{
This function is amazing and beautiful
}
\examples{
redplot(x=iris$Petal.Length, y=iris$Petal.Width)
}
