% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{This says hello to any person!}
\usage{
hello(name = "Andreas")
}
\arguments{
\item{name}{A string of characters. Preferentiallly a name of a person. (Default="Andreas")}
}
\value{
Prints Hello name to screen
}
\description{
Here jkhfjklsdfghdfjklhparam
}
\examples{
hello("Malie")
}
