% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Says hello to anyone.}
\usage{
hello(name = "Malie")
}
\arguments{
\item{name}{A string of characters. Preferably a name of a person. (Default="Malie")}
}
\value{
Prints Hello name to screen.
}
\description{
Says hello to anyone.
}
\examples{
hello("Andreas")
}
